+++ b2/utils.go
@@ -1,0 +1,18 @@
+// ####################################################################
+//
+// File: b2/utils.go
+//
+// Copyright 2024 Backblaze Inc. All Rights Reserved.
+//
+// License https://www.backblaze.com/using_b2_code.html
+//
+// ####################################################################
+
+package b2
+
+func If[T any](cond bool, vtrue, vfalse T) T {
+	if cond {
+		return vtrue
+	}
+	return vfalse
+}